fix: BottomSheet types [MDS-357]
