+++ workspaces/core/src/bottomSheet/private/types/BackdropProps.ts
@@ -1,0 +1,5 @@
+type BackdropProps = {
+  className?: string;
+};
+
+export default BackdropProps;
