+++ workspaces/core/src/bottomSheet/private/types/BottomSheetRootProps.ts
@@ -1,0 +1,11 @@
+import type Size from './Size';
+
+type BottomSheetRootProps = {
+  open: boolean;
+  onClose: () => void;
+  hasShadow?: boolean;
+  size?: Size;
+  children?: React.ReactNode;
+};
+
+export default BottomSheetRootProps;
