+++ workspaces/core/src/bottomSheet/private/types/PanelProps.ts
@@ -1,0 +1,8 @@
+type PanelProps = {
+  className?: string;
+  hasShadow?: boolean;
+  onClose?: () => void;
+  children?: React.ReactNode;
+};
+
+export default PanelProps;
