+++ workspaces/core/src/bottomSheet/private/types/Size.ts
@@ -1,0 +1,3 @@
+type Size = 'sm' | 'md' | 'lg';
+
+export default Size;
